airflow/requirements.txt
kafka-python==2.0.2
clickhouse-driver==0.2.6
clickhouse-connect==0.6.23
pandas==2.1.3
jsonschema==4.20.0
